refactor(slack): modularize adapter and implement security detector
- Split monolithic slack adapter into specialized sub-modules
- Enhance native assistant streaming and status support
- Optimize package documentation and add README.md
- Implement system-wide security detector (WAF) for command interception
- Add .serena configuration files
